Create functionality to active write a text note This functionality allow the user when the modal to create a new note is open and the user click on the option to add a new note by texting, will open editor (textarea componet) to start write a text note.
This functionality is possible because of useState hook from react, that allows to change state variable value. So we created a state to control if we should show the textarea componet or not.

We also create a function that change the value to false or true depending on the users click